Bug fix castling and Checks
sound s when moving and when capturing are now being played correctly. castling is now on both sides available when neither the king or the corresponding rook hasn't moved yet and when a piece in pined to your king it is not able to move anymore. this has some minor bug but it works for now. also when you are in check all moves except moves that get you out of check are now valid